use [u]intptr_t instead of [unsigned] int for pointers

+++ vm/runtime/uncommonBranch.cpp
@@ -38,18 +38,18 @@ bool patch_uncommon_call(frame* f) {
 
   int* next_inst = (int*) f->pc();
   int* dest_addr = next_inst-1;
-  int  dest      = *dest_addr + (int) next_inst;
+  int  dest      = *dest_addr + (intptr_t) next_inst;
 
   // return true if the call has been executed before
-  if (dest == (int)StubRoutines::used_uncommon_trap_entry())
+  if (dest == (intptr_t)StubRoutines::used_uncommon_trap_entry())
     return true;
 
-  assert(dest == (int)StubRoutines::unused_uncommon_trap_entry(), "Make sure we are patching the right call");
+  assert(dest == (intptr_t)StubRoutines::unused_uncommon_trap_entry(), "Make sure we are patching the right call");
 
   // patch with used_uncommon_trap
-  *dest_addr = ((int)StubRoutines::used_uncommon_trap_entry()) - ((int) next_inst);
+  *dest_addr = ((intptr_t)StubRoutines::used_uncommon_trap_entry()) - ((intptr_t) next_inst);
 
-  assert(*dest_addr + (int) next_inst == (int)StubRoutines::used_uncommon_trap_entry(), "Check the patch");
+  assert(*dest_addr + (intptr_t) next_inst == (intptr_t)StubRoutines::used_uncommon_trap_entry(), "Check the patch");
 
   // return false since the call is patched
   return false;
